--- a/app/src/app-pch.h
+++ b/app/src/app-pch.h
@@ -1,0 +1,9 @@
+/* Add C includes here */
+
+#if defined __cplusplus
+/* Add C++ includes here */
+
+#include <QObject>
+#include <QString>
+#include <QWidget>
+#endif