--- a/core_lib/src/canvaspainter.cpp
+++ b/core_lib/src/canvaspainter.cpp
@@ -16,6 +16,7 @@ GNU General Public License for more details.
 
 #include "canvaspainter.h"
 
+#include <QtMath>
 #include "object.h"
 #include "layerbitmap.h"
 #include "layervector.h"