--- a/core_lib/src/corelib-pch.h
+++ b/core_lib/src/corelib-pch.h
@@ -1,0 +1,10 @@
+
+/* Add C includes here */
+
+#if defined __cplusplus
+/* Add C++ includes here */
+
+#include <QObject>
+#include <QString>
+#include <QList>
+#endif