--- a/core_lib/src/graphics/vector/beziercurve.h
+++ b/core_lib/src/graphics/vector/beziercurve.h
@@ -16,11 +16,12 @@ GNU General Public License for more details.
 #ifndef BEZIERCURVE_H
 #define BEZIERCURVE_H
 
-#include <QtXml>
 #include <QPainter>
 
 class Object;
 class Status;
+class QXmlStreamWriter;
+class QDomElement;
 
 struct Intersection
 {
@@ -37,7 +38,7 @@ public:
     explicit BezierCurve(const QList<QPointF>& pointList, const QList<qreal>& pressureList, double tol, bool smooth=true);
 
     Status createDomElement(QXmlStreamWriter &xmlStream);
-    void loadDomElement(QDomElement element);
+    void loadDomElement(const QDomElement& element);
 
     qreal getWidth() const { return width; }
     qreal getFeather() const { return feather; }