--- a/core_lib/src/graphics/vector/vectorimage.cpp
+++ b/core_lib/src/graphics/vector/vectorimage.cpp
@@ -18,6 +18,10 @@ GNU General Public License for more details.
 
 #include <cmath>
 #include <QImage>
+#include <QFile>
+#include <QFileInfo>
+#include <QDebug>
+#include <QXmlStreamWriter>
 #include "object.h"