--- a/core_lib/src/interface/timecontrols.cpp
+++ b/core_lib/src/interface/timecontrols.cpp
@@ -17,8 +17,8 @@ GNU General Public License for more details.
 
 #include "timecontrols.h"
 
-#include <QtGui>
 #include <QLabel>
+#include <QSettings>
 
 #include "editor.h"
 #include "playbackmanager.h"