Precompiled headers can reduce build time significantly. 

In this PR, I tried:
1. Add precompiled headers to app & core_lib
2. Minimise the header includes in .h files
3. Use forward declaration as much as possible

On my laptop, the build time of a Release full rebuild goes down from 167 sec to 75 sec.

It also helps our CI:
- Travis CI Linux: 5 min 38 sec -> 4 min 43 sec
- Travis CI Mac: 7 min 33 sec -> 5 min 15 sec
- AppVeyor Win32: 12 min 23 sec ->  8 min 13 sec
- AppVeyor Win64: 13 min 16 sec -> 9 min 49 sec